Update client build
