+++ package.json
@@ -20,44 +20,25 @@
     "url": "https://github.com/WNPRC-EHR-Services/wnprc-modules/issues"
   },
   "homepage": "https://github.com/WNPRC-EHR-Services/wnprc-modules#readme",
-  "devDependencies": {
-    "@types/node": "^7.0.64",
-    "@types/qunit": "^2.5.0",
-    "@types/react": "^15.6.15",
-    "@types/react-dom": "^15.5.7",
-    "@types/rsvp": "~3.3.0",
-    "@types/toastr": "^2.0.0",
-    "@types/uuid": "^3.0.0",
-    "@types/webpack": "^3.8.12",
-    "awesome-typescript-loader": "^3.5.0",
-    "expose-loader": "^0.7.5",
-    "less": "^3.0.4",
-    "ts-loader": "^3.5.0",
-    "ts-node": "^3.3.0",
-    "tslint": "^5.10.0",
-    "tslint-react": "^3.6.0",
-    "typescript": "~2.4.0",
-    "webpack": "^3.12.0"
-  },
   "dependencies": {
     "@wnprc/primate-id": "^1.1.7",
-    "blockui": "^1.0.0",
-    "bootstrap": "^3.0.0",
+    "blockui": "1.0.0",
+    "bootstrap": "3.4.1",
     "fullcalendar": "^3.0.0",
-    "jquery": "^3.5.0",
+    "jquery": "3.5.1",
     "knockout": "3.5.0",
     "knockout.mapping": "~2.4.0",
     "knockout.punches": "~0.5.0",
     "lodash": "^4.17.13",
     "markdown": "^0.5.0",
-    "moment": "^2.0.0",
+    "moment": "2.29.1",
     "qunit": "^2.6.1",
-    "react": "^15.0.0",
-    "react-bootstrap-typeahead": "^2.0.0-rc.4",
-    "react-data-grid": "^2.0.78",
-    "react-data-grid-addons": "^2.0.78",
-    "react-dom": "^15.0.0",
-    "react-tabs": "^1.0.0",
+    "react": "16.14.0",
+    "react-bootstrap-typeahead": "2.6.0",
+    "react-data-grid": "2.0.80",
+    "react-data-grid-addons": "2.0.80",
+    "react-dom": "16.14.0",
+    "react-tabs": "1.1.0",
     "rsvp": "~3.3.0",
     "simplefilter": "^1.0.0",
     "toastr": "^2.0.0",
@@ -66,6 +47,21 @@
     "urijs": "^1.0.0",
     "uuid": "^3.0.0",
     "whatwg-fetch": "^2.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^7.0.64",
+    "@types/qunit": "^2.5.0",
+    "@types/react": "16.14.3",
+    "@types/react-dom": "16.9.10",
+    "@types/rsvp": "~3.3.0",
+    "@types/toastr": "^2.0.0",
+    "@types/uuid": "^3.0.0",
+    "less": "3.13.1",
+    "source-map-loader": "1.1.3",
+    "ts-loader": "8.0.17",
+    "typescript": "4.0.3",
+    "webpack": "4.46.0",
+    "webpack-cli": "4.5.0"
   },
   "optionalDependencies": {
     "fsevents": "^1.2.4"
