+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "target": "es6",
+    "jsx": "react",
+    "sourceMap": false,
+    "moduleResolution": "node",
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "types": [] // don't implicitly include any @types, only those included via import
+  },
+  "exclude": [
+    "node_modules",
+    "resources"
+  ]
+}
